✨ feat: Finish video swiper player component encapsulation
